requirements.txt
asgiref==3.5.0
Django==4.0.2
django-crispy-forms==1.14.0
django-filter==21.1
mysqlclient==2.1.0
pytz==2021.3
sqlparse==0.4.2
